refac: rename game situation check function
